+++ inc/Providers/OpenAI.php
@@ -11,6 +11,7 @@
 namespace AiPlusBlockEditor\Providers;
 
 use Orhanerday\OpenAi\OpenAi as ChatGPT;
+use AiPlusBlockEditor\Admin\Options;
 use AiPlusBlockEditor\Interfaces\Provider;
 
 class OpenAI implements Provider {
@@ -44,6 +45,21 @@ class OpenAI implements Provider {
 	}
 
 	/**
+	 * Get Client.
+	 *
+	 * This method initializes the OpenAI client
+	 * with the API keys stored in the options.
+	 *
+	 * @since 1.5.0
+	 *
+	 * @return ChatGPT
+	 */
+	protected function get_client(): ChatGPT {
+		$ai_keys = get_option( Options::get_page_option(), [] )['open_ai_token'] ?? '';
+		return new ChatGPT( $ai_keys );
+	}
+
+	/**
 	 * Get AI Response.
 	 *
 	 * @since 1.0.0
@@ -52,11 +68,11 @@ class OpenAI implements Provider {
 	 * @return string|\WP_Error
 	 */
 	public function run( $payload ) {
-		$ai_keys = get_option( 'ai_plus_block_editor', [] )['open_ai_token'] ?? '';
+		$ai_keys = get_option( Options::get_page_option(), [] )['open_ai_token'] ?? '';
 		$payload = wp_parse_args( [ 'role' => 'user' ], $payload );
 
 		try {
-			$response = ( new ChatGPT( $ai_keys ) )->chat(
+			$response = $this->get_client()->chat(
 				wp_parse_args(
 					[ 'messages' => [ $payload ] ],
 					$this->get_default_args()
@@ -67,19 +83,26 @@ class OpenAI implements Provider {
 
 			// Deal gracefully, with API error.
 			if ( isset( $response['error'] ) ) {
-				return $this->get_json_error( $response['error']['message'] ?? '' );
+				$error_msg = $response['error']['message'] ?? '';
+				error_log( $error_msg );
+
+				return $this->get_json_error( $error_msg );
 			}
 		} catch ( \Exception $e ) {
 			$error_msg = sprintf(
 				'Error: OpenAI API call failed... %s',
 				$e->getMessage()
 			);
+			error_log( $error_msg );
 
 			return $this->get_json_error( $error_msg );
 		}
 
 		if ( is_null( $response ) ) {
-			return $this->get_json_error( 'Error: Malformed JSON output.' );
+			$error_msg = 'Error: OpenAI API call returned malformed JSON.';
+			error_log( $error_msg );
+
+			return $this->get_json_error( $error_msg );
 		}
 
 		return $response['choices'][0]['message']['content'] ?? '';