+++ src/components/index.tsx
@@ -1,0 +1,8 @@
+import SEO from '../components/SEO';
+import Slug from '../components/Slug';
+import Summary from '../components/Summary';
+import Headline from '../components/Headline';
+import Switcher from '../components/Switcher';
+import Social from '../components/Social';
+
+export { SEO, Slug, Summary, Headline, Switcher, Social };