+++ tests/php/Services/BootTest.php
@@ -11,6 +11,12 @@ use AiPlusBlockEditor\Abstracts\Service;
  * @covers \AiPlusBlockEditor\Services\Boot::register
  * @covers \AiPlusBlockEditor\Services\Boot::register_translation
  * @covers \AiPlusBlockEditor\Services\Boot::register_scripts
+ * @covers \AiPlusBlockEditor\Admin\Options::__callStatic
+ * @covers \AiPlusBlockEditor\Admin\Options::get_form_fields
+ * @covers \AiPlusBlockEditor\Admin\Options::get_form_notice
+ * @covers \AiPlusBlockEditor\Admin\Options::get_form_page
+ * @covers \AiPlusBlockEditor\Admin\Options::get_form_submit
+ * @covers \AiPlusBlockEditor\Admin\Options::init
  */
 class BootTest extends TestCase {
 	public Boot $boot;
@@ -36,6 +42,27 @@ class BootTest extends TestCase {
 
 	public function test_register_scripts() {
 		$boot = new \ReflectionClass( Boot::class );
+
+		$mock_boot = Mockery::mock( Boot::class )->makePartial();
+		$mock_boot->shouldAllowMockingProtectedMethods();
+
+		$mock_boot->shouldReceive( 'get_assets' )
+			->andReturn(
+				[
+					'dependencies' => [
+						'wp-i18n',
+						'wp-element',
+						'wp-blocks',
+						'wp-components',
+						'wp-editor',
+						'wp-hooks',
+						'wp-compose',
+						'wp-plugins',
+						'wp-edit-post',
+					],
+					'version'      => '1750321560',
+				]
+			);
 
 		\WP_Mock::userFunction( 'plugins_url' )
 			->andReturnUsing(
@@ -63,9 +90,33 @@ class BootTest extends TestCase {
 					'wp-plugins',
 					'wp-edit-post',
 				],
-				'1.3.0',
+				'1750321560',
 				false,
 			);
+
+		\WP_Mock::userFunction( 'get_option' )
+			->andReturnUsing(
+				function ( $arg ) {
+					return $arg;
+				}
+			);
+
+		\WP_Mock::userFunction( 'esc_html__' )
+			->andReturnUsing(
+				function ( $arg ) {
+					return $arg;
+				}
+			);
+
+		\WP_Mock::userFunction( 'esc_attr' )
+			->andReturnUsing(
+				function ( $arg ) {
+					return $arg;
+				}
+			);
+
+		\WP_Mock::userFunction( 'wp_localize_script' )
+			->andReturn( null );
 
 		\WP_Mock::userFunction( 'wp_set_script_translations' )
 			->with(
@@ -74,13 +125,27 @@ class BootTest extends TestCase {
 				'/var/www/wp-content/plugins/ai-plus-block-editor/inc/Services/../../languages',
 			);
 
-		$this->boot->register_scripts();
+		$mock_boot->register_scripts();
 
 		$this->assertConditionsMet();
 	}
 
 	public function test_register_translation() {
 		$boot = new \ReflectionClass( Boot::class );
+
+		\WP_Mock::userFunction( 'esc_html__' )
+			->andReturnUsing(
+				function ( $arg ) {
+					return $arg;
+				}
+			);
+
+		\WP_Mock::userFunction( 'esc_attr' )
+			->andReturnUsing(
+				function ( $arg ) {
+					return $arg;
+				}
+			);
 
 		\WP_Mock::userFunction( 'plugin_basename' )
 			->once()