+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+# 1.5.0
+* Feat: Add AI Provider Switcher.
+* Feat: Add Social Media Hashtags Feature.
+* Fix: Breaking issue with Tone functionality.
+* Test: Fix Unit tests & update snapshots.
+* Chore: Relocate `apbe.blockMenuOptions` hook.
+* Feat: Update text translations.
+
 # 1.4.0
 * Refactor: Adopt useSelect & useDispatch hooks in sidebar components.
 * Feat: Add new language translations for Italian, Russian, Chinese, Arabic, Hebrew & Croatian.